Start generalizing VM unit tests
