--- core/lib/multivm/src/versions/vm_fast/tests/tester/mod.rs
@@ -1,6 +1,0 @@
-pub(crate) use transaction_test_info::{ExpectedError, TransactionTestInfo, TxModifier};
-pub(crate) use vm_tester::{get_empty_storage, VmTester, VmTesterBuilder};
-pub(crate) use zksync_test_account::{Account, DeployContractsTx, TxType};
-
-mod transaction_test_info;
-mod vm_tester;
